refactor code to pass props for conversation key instead of using routes
