+++ apps/web/src/components/Messages/Push/PUSHPreview.tsx
@@ -1,44 +1,23 @@
-import { chat, chats } from '@components/utils/hooks/push/data';
 import useCreateChatProfile from '@components/utils/hooks/push/useCreateChatProfile';
 // import useFetchChat from '@components/utils/hooks/push/useFetchChat';
 import { Trans } from '@lingui/macro';
-import { useRouter } from 'next/router';
-import { type FC, useEffect, useState } from 'react';
+import { type FC } from 'react';
 import { PUSH_TABS, usePushChatStore } from 'src/store/push-chat';
 import { Card, Input, Modal } from 'ui';
+
+import PUSHPreviewChats from './PUSHPreviewChats';
+import PUSHPreviewRequests from './PUSHPreviewRequests';
 
 interface PreviewListProps {
   selectedConversationKey?: string;
 }
 
-const PUSHPreview: FC<PreviewListProps> = () => {
-  const router = useRouter();
-
+const PUSHPreview: FC<PreviewListProps> = ({ selectedConversationKey }) => {
   const activeTab = usePushChatStore((state) => state.activeTab);
   const setActiveTab = usePushChatStore((state) => state.setActiveTab);
   const showCreateChatProfileModal = usePushChatStore((state) => state.showCreateChatProfileModal);
   const setShowCreateChatProfileModal = usePushChatStore((state) => state.setShowCreateChatProfileModal);
   const { modalContent, isModalClosable } = useCreateChatProfile();
-  // todo once we have the type of chats, replace it with 'any' here
-  const [allChats, setAllChats] = useState<any>([]);
-
-  useEffect(() => {
-    // const chats = useFetchChats()
-    // todo once we have the type of chats, replace it with 'any' here
-    const userChats = [...chats].map((oneChat: any) => {
-      // populate the latest chat for this conversation
-      // const response = useFetchChat()
-      oneChat.chat = chat;
-      return oneChat;
-    });
-    setAllChats(userChats);
-  }, []);
-
-  const [conversationId] = router.query.conversationKey as string[];
-
-  const clickChat = (key: string) => {
-    router.push(`/messages/push/${key}`);
-  };
 
   return (
     <div className="flex h-full flex-col justify-between">
@@ -79,33 +58,11 @@ const PUSHPreview: FC<PreviewListProps> = () => {
         {/* section for header */}
         {/* section for chats */}
         {activeTab === PUSH_TABS.CHATS && (
-          <section className="flex flex-col	gap-2.5	">
-            {/* todo change the type here from 'any' to another typw */}
-            {allChats.map((oneChat: any, number: number) => (
-              <div
-                key={number}
-                className={`flex h-16 cursor-pointer gap-2.5 rounded-lg  p-2.5 pr-3 transition-all hover:bg-gray-100 ${
-                  conversationId === oneChat.id && 'bg-violet-100'
-                }`}
-                onClick={() => clickChat(oneChat.id)}
-              >
-                <img className="h-12	w-12 rounded-full" src="/user.svg" alt="" />
-                <div className="flex w-full	justify-between	">
-                  <div>
-                    <p className="bold text-base">{oneChat.name}</p>
-                    <p className="text-sm text-gray-500	">{oneChat.chat.text}</p>
-                  </div>
-                  <div>
-                    <span className="text-xs text-gray-500">{oneChat.chat.time}</span>
-                  </div>
-                </div>
-              </div>
-            ))}
-          </section>
+          <PUSHPreviewChats selectedConversationKey={String(selectedConversationKey)} />
         )}
         {/* section for chats */}
         {/* sections for requests */}
-        {activeTab === PUSH_TABS.REQUESTS && <section>requests</section>}
+        {activeTab === PUSH_TABS.REQUESTS && <PUSHPreviewRequests />}
         {/* sections for requests */}
       </Card>
       {/* <button onClick={createChatProfile}>Create Profile</button> */}
