+++ apps/web/src/components/Messages/Push/PUSHPreviewRequests.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function PUSHPreviewRequests() {
+  // const router = useRouter();
+  // const [allChats, setAllChat] = useState<String[]>([]);
+
+  // const clickChat = (key: string) => {
+  // router.push(`/messages/push/${key}`);
+  // };
+
+  return <section className="flex flex-col	gap-2.5	">All requests</section>;
+}
